+++ packages/js-utils/src/index.ts
@@ -33,6 +33,8 @@ export {
   Units
 }
 
+
+export * from './types'
 export type {
   CommonLogger
 }