+++ packages/js-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @goatlab/js-utils
 
+## 0.8.8
+
+### Patch Changes
+
+- Add Object methods
+
+## 0.8.7
+
+### Patch Changes
+
+- Export types
+
 ## 0.8.6
 
 ### Patch Changes